Fixes #493: compatibility with plugin "Enable Media Replace" broken with WP 5.3 (PR #497)
* Use the hook 'wp_handle_replace' to initiate the class
Also removed some `if` statements and tests that are not needed anymore, AMR has already done them.
* Import more classes
* Deprecate method store_old_backup_path()
- The old webp file paths are stored in `init()` now: like explained in the DocBlock, this prevents some orphan files from being kept indefinitely when the user chooses not to rename the files.
- The method `store_old_backup_path()` is not needed anymore: the old backup path is already correct and the old webp file paths are stored in `init()`.
* Removed @access and @author tags, imported trait
* Removed useless "if" statement
* Improve comments
* Fixed auto-optimization for existing media
The new argument `$context` passed to the filter `wp_generate_attachment_metadata` cannot be trusted. This argument’s value was meant to be `'create'`for new uploads and `'update'` for the other cases. In reality, it can be `'create'` when the media is not a new upload: this is the value used in the function `wp_generate_attachment_metadata()`, but this function is used by developers to re-create the thumbnails, this doesn't mean it is a new upload!
* Made the class Imagify_Auto_Optimization_Deprecated abstract