Re-namespace and start add_2x_sizes filter method
Re-namespace and change filenames to bring the compatibility classes more in line with PS4 in preparation for adding unit tests. Adjust composer to point namespace to proper directory.

Create a method to filter media sizes, where we can add @2x sizes into into the files array of a media we are about to optimize. (Starts by just returning the given files array.)