Add re-optimization of images after regeneration
Adds a method to restore a full-size image from the backed up original.
Adds a method to replace back the previously optimized full-size image after image regeneration.
Adds a method to send regenerated images to the queue for re-optimization.

We also hook the restore/replace before and after the generation of retina sizes to insure that the original is being used to generate the retinas as well.