Check for new upload before doing retina process
For new uploads, we're going optimize everything after all the other things run their stuff.
In this case we don't want to start a proces for this media now -- doing so will "Lock" the media when the full process tries to run in a few moments, and while retinas will be processed nothing else will be included!