PRevent Fatal Error on server
Actually, when update version 2.0 to 2.1 on a PHP 7.0 server, a fatal error occure. 

The error: `Fatal error: Composer detected issues in your platform: Your Composer dependencies require a PHP version ">= 7.2.0". You are running 7.0.33-0+deb9u12. in wp-content/plugins/imagify/vendor/composer/platform_check.php on line 24`

Wordpress PHP version in the plugin file need to be changed in order to not give the user the right to download the plugin.